chore: Create dedicated helper function for generating list of dates used for recap range
